chore: reduce circular deps in Table utils
