+++ packages/grafana-ui/src/internal/index.ts
@@ -68,7 +68,7 @@ export {
 export { defaultSparklineCellConfig } from '../components/Table/Cells/SparklineCell';
 export { TableCell } from '../components/Table/Cells/TableCell';
 export { useTableStyles } from '../components/Table/TableRT/styles';
-export { migrateTableDisplayModeToCellOptions } from '../components/Table/utils';
+export { migrateTableDisplayModeToCellOptions } from '../components/Table/cellUtils';
 export { type DataLinksContextMenuApi } from '../components/DataLinks/DataLinksContextMenu';
 export { MenuDivider } from '../components/Menu/MenuDivider';
 export { AbstractList } from '../components/List/AbstractList';
